refactor: 퀴즈 결과 응답 조립 분리
