+++ src/test/java/com/f1/quiket/domain/quiz/service/QuizResultSubmitServiceTest.java
@@ -77,7 +77,8 @@ class QuizResultSubmitServiceTest {
                 quizPlayAnswerRepository,
                 quizResultRepository,
                 userRepository,
-                gamificationRewardService
+                gamificationRewardService,
+                new QuizResultResponseAssembler()
         );
     }
